# docs/openapi.yaml
{
  "openapi":"3.0.1",
  "info":{
    "title":"OpenAPI definition",
    "version":"v0"
  },
  "servers":[
    {
      "url":"http://localhost:8080",
      "description":"Generated server url"
    }
  ],
  "paths":{
    "/users":{
      "get":{
        "tags":[
          "user-controller"
        ],
        "operationId":"getAllBankCard",
        "responses":{
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "type":"array",
                  "items":{
                    "$ref":"#/components/schemas/UserDto"
                  }
                }
              }
            }
          }
        }
      },
      "post":{
        "tags":[
          "user-controller"
        ],
        "operationId":"createUser",
        "requestBody":{
          "content":{
            "application/json":{
              "schema":{
                "$ref":"#/components/schemas/RegistrationUserDto"
              }
            }
          },
          "required":true
        },
        "responses":{
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/UserDto"
                }
              }
            }
          }
        }
      }
    },
    "/transfer":{
      "post":{
        "tags":[
          "transfer-controller"
        ],
        "operationId":"bankCardTransfer",
        "requestBody":{
          "content":{
            "application/json":{
              "schema":{
                "$ref":"#/components/schemas/BankTransferDto"
              }
            }
          },
          "required":true
        },
        "responses":{
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "type":"array",
                  "items":{
                    "$ref":"#/components/schemas/BankCardUpdateDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/cards":{
      "get":{
        "tags":[
          "bank-card-controller"
        ],
        "operationId":"getCards",
        "parameters":[
          {
            "name":"pageable",
            "in":"query",
            "required":true,
            "schema":{
              "$ref":"#/components/schemas/Pageable"
            }
          }
        ],
        "responses":{
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/PageBankCardDto"
                }
              }
            }
          }
        }
      },
      "post":{
        "tags":[
          "bank-card-controller"
        ],
        "operationId":"createBankCard",
        "requestBody":{
          "content":{
            "application/json":{
              "schema":{
                "$ref":"#/components/schemas/BankCardDto"
              }
            }
          },
          "required":true
        },
        "responses":{
          "200":{
            "description":"OK",
            "content":{
              "application/json":{
                "schema":{
                  "type":"object"
                }
              }
            }
          }
        }
      }
    },
    "/auth":{
      "post":{
        "tags":[
          "auth-controller"
        ],
        "operationId":"createAuthToken",
        "requestBody":{
          "content":{
            "application/json":{
              "schema":{
                "$ref":"#/components/schemas/JwtRequest"
              }
            }
          },
          "required":true
        },
        "responses":{
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "type":"object"
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}":{
      "delete":{
        "tags":[
          "user-controller"
        ],
        "operationId":"deleteBankCard",
        "parameters":[
          {
            "name":"id",
            "in":"path",
            "required":true,
            "schema":{
              "type":"integer",
              "format":"int64"
            }
          }
        ],
        "responses":{
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/UserDto"
                }
              }
            }
          }
        }
      },
      "patch":{
        "tags":[
          "user-controller"
        ],
        "operationId":"updateUser",
        "parameters":[
          {
            "name":"id",
            "in":"path",
            "required":true,
            "schema":{
              "type":"integer",
              "format":"int64"
            }
          }
        ],
        "requestBody":{
          "content":{
            "application/json":{
              "schema":{
                "$ref":"#/components/schemas/UpdateUserDto"
              }
            }
          },
          "required":true
        },
        "responses":{
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/UserDto"
                }
              }
            }
          }
        }
      }
    },
    "/cards/{id}":{
      "delete":{
        "tags":[
          "bank-card-controller"
        ],
        "operationId":"deleteBankCard_1",
        "parameters":[
          {
            "name":"id",
            "in":"path",
            "required":true,
            "schema":{
              "type":"integer",
              "format":"int64"
            }
          }
        ],
        "responses":{
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/BankCardDto"
                }
              }
            }
          }
        }
      },
      "patch":{
        "tags":[
          "bank-card-controller"
        ],
        "operationId":"patchBankCard",
        "parameters":[
          {
            "name":"id",
            "in":"path",
            "required":true,
            "schema":{
              "type":"integer",
              "format":"int64"
            }
          }
        ],
        "requestBody":{
          "content":{
            "application/json":{
              "schema":{
                "$ref":"#/components/schemas/BankCardUpdateDto"
              }
            }
          },
          "required":true
        },
        "responses":{
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/BankCardDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components":{
    "schemas":{
      "RegistrationUserDto":{
        "required":[
          "cardHolderName",
          "email",
          "password",
          "passwordConfirm"
        ],
        "type":"object",
        "properties":{
          "cardHolderName":{
            "type":"string"
          },
          "password":{
            "type":"string"
          },
          "passwordConfirm":{
            "type":"string"
          },
          "email":{
            "type":"string"
          }
        }
      },
      "UserDto":{
        "required":[
          "cardHolderName",
          "email"
        ],
        "type":"object",
        "properties":{
          "cardHolderName":{
            "type":"string"
          },
          "email":{
            "type":"string"
          }
        }
      },
      "BankTransferDto":{
        "required":[
          "recipientsCard",
          "sendersCard",
          "summa"
        ],
        "type":"object",
        "properties":{
          "sendersCard":{
            "maxLength":2147483647,
            "minLength":12,
            "pattern":"\\d+",
            "type":"string"
          },
          "recipientsCard":{
            "maxLength":2147483647,
            "minLength":12,
            "pattern":"\\d+",
            "type":"string"
          },
          "summa":{
            "type":"integer",
            "format":"int64"
          }
        }
      },
      "BankCardUpdateDto":{
        "type":"object",
        "properties":{
          "userDto":{
            "$ref":"#/components/schemas/UserDto"
          },
          "validPeriod":{
            "type":"string",
            "format":"date-time"
          },
          "balance":{
            "type":"integer",
            "format":"int64"
          },
          "statusCard":{
            "type":"string",
            "enum":[
              "ACTIVE",
              "INACTIVE",
              "DEADLINE_EXPIRED"
            ]
          }
        }
      },
      "BankCardDto":{
        "required":[
          "balance",
          "statusCard",
          "validPeriod"
        ],
        "type":"object",
        "properties":{
          "userDto":{
            "$ref":"#/components/schemas/UserDto"
          },
          "validPeriod":{
            "type":"string",
            "format":"date-time"
          },
          "balance":{
            "type":"integer",
            "format":"int64"
          },
          "bankCardNumber":{
            "type":"string"
          },
          "statusCard":{
            "type":"string",
            "enum":[
              "ACTIVE",
              "INACTIVE",
              "DEADLINE_EXPIRED"
            ]
          }
        }
      },
      "JwtRequest":{
        "required":[
          "cardHolderName",
          "password"
        ],
        "type":"object",
        "properties":{
          "cardHolderName":{
            "type":"string"
          },
          "password":{
            "type":"string"
          }
        }
      },
      "RoleDto":{
        "required":[
          "roleName"
        ],
        "type":"object",
        "properties":{
          "roleName":{
            "type":"string"
          }
        }
      },
      "UpdateUserDto":{
        "type":"object",
        "properties":{
          "cardHolderName":{
            "type":"string"
          },
          "roleDtoCollection":{
            "type":"array",
            "items":{
              "$ref":"#/components/schemas/RoleDto"
            }
          },
          "bankCardDtoList":{
            "type":"array",
            "items":{
              "$ref":"#/components/schemas/BankCardDto"
            }
          },
          "email":{
            "type":"string"
          }
        }
      },
      "Pageable":{
        "type":"object",
        "properties":{
          "page":{
            "minimum":0,
            "type":"integer",
            "format":"int32"
          },
          "size":{
            "minimum":1,
            "type":"integer",
            "format":"int32"
          },
          "sort":{
            "type":"array",
            "items":{
              "type":"string"
            }
          }
        }
      },
      "PageBankCardDto":{
        "type":"object",
        "properties":{
          "totalElements":{
            "type":"integer",
            "format":"int64"
          },
          "totalPages":{
            "type":"integer",
            "format":"int32"
          },
          "size":{
            "type":"integer",
            "format":"int32"
          },
          "content":{
            "type":"array",
            "items":{
              "$ref":"#/components/schemas/BankCardDto"
            }
          },
          "number":{
            "type":"integer",
            "format":"int32"
          },
          "sort":{
            "$ref":"#/components/schemas/SortObject"
          },
          "first":{
            "type":"boolean"
          },
          "last":{
            "type":"boolean"
          },
          "numberOfElements":{
            "type":"integer",
            "format":"int32"
          },
          "pageable":{
            "$ref":"#/components/schemas/PageableObject"
          },
          "empty":{
            "type":"boolean"
          }
        }
      },
      "PageableObject":{
        "type":"object",
        "properties":{
          "offset":{
            "type":"integer",
            "format":"int64"
          },
          "sort":{
            "$ref":"#/components/schemas/SortObject"
          },
          "pageSize":{
            "type":"integer",
            "format":"int32"
          },
          "pageNumber":{
            "type":"integer",
            "format":"int32"
          },
          "paged":{
            "type":"boolean"
          },
          "unpaged":{
            "type":"boolean"
          }
        }
      },
      "SortObject":{
        "type":"object",
        "properties":{
          "empty":{
            "type":"boolean"
          },
          "sorted":{
            "type":"boolean"
          },
          "unsorted":{
            "type":"boolean"
          }
        }
      }
    }
  }
}